admin_dashboard/requirements.txt
streamlit
folium
streamlit-folium
google-generativeai
pillow
firebase-admin
google-auth
